+++ product_portfolio/models.py
@@ -251,8 +251,8 @@ class Product(BaseProductMixin, FieldChangesMixin, KeywordsMixin, DataspacedMode
     def get_check_package_version_url(self):
         return self.get_url("check_package_version")
 
-    def get_import_manifest_url(self):
-        return self.get_url("import_manifest")
+    def get_load_sboms_url(self):
+        return self.get_url("load_sboms")
 
     def get_pull_project_data_url(self):
         return self.get_url("pull_project_data")
@@ -1119,7 +1119,9 @@ class ScanCodeProject(HistoryFieldsMixin, DataspacedModel):
     """Wrap a ScanCode.io Project."""
 
     class ProjectType(models.TextChoices):
+        # This type was replaced by LOAD_SBOMS but is kept for backward compatibility
         IMPORT_FROM_MANIFEST = "IMPORT_FROM_MANIFEST", _("Import from Manifest")
+        LOAD_SBOMS = "LOAD_SBOMS", _("Load SBOMs")
         PULL_FROM_SCANCODEIO = "PULL_FROM_SCANCODEIO", _("Pull from ScanCode.io")
 
     class Status(models.TextChoices):